refactor: déplacer authClient vers utils et corriger les chemins d'importation
