+++ pages/index.vue
@@ -1,7 +1,11 @@
 <script setup lang="ts">
-import { authClient } from '~/server/utils/auth-client'
+import { authClient } from '~/utils/auth-client'
 
 const session = authClient.useSession()
+
+useHead({
+  title: 'Le Coffre',
+})
 </script>
 
 <template>
